Split up [github] testers
