+++ services/github/github-repo-size.tester.js
@@ -1,0 +1,22 @@
+'use strict'
+
+const Joi = require('joi')
+const { isFileSize } = require('../test-validators')
+
+const t = (module.exports = require('../create-service-tester')())
+
+t.create('repository size')
+  .get('/badges/shields.json')
+  .expectJSONTypes(
+    Joi.object().keys({
+      name: 'repo size',
+      value: isFileSize,
+    })
+  )
+
+t.create('repository size (repo not found)')
+  .get('/badges/helmets.json')
+  .expectJSON({
+    name: 'repo size',
+    value: 'repo not found',
+  })
